+++ src/modules/translation/types.ts
@@ -4,3 +4,8 @@ export interface Translation {
 export interface TranslationKeys {
   [key: string]: string
 }
+
+export type TranslationSagaOptions = {
+  getTranslation?: (locale: string) => Promise<Translation>
+  translations?: { [locale: string]: Translation }
+}
